FOUN-1374: fix numpy int regression

+++ biophi/humanization/methods/humanness.py
@@ -79,7 +79,7 @@ class ChainHumanness:
         if frequency:
             curve = curve / curve.sum()
         else:
-            curve = curve.astype(np.int)
+            curve = curve.astype(np.int32)
 
         if cumulative:
             curve = curve[::-1].cumsum()[::-1]
